Add meal dialog to screen and layout fixes
