+++ app/src/main/java/com/dj/insulink/feature/meals/ui/DateTimeInput.kt
@@ -1,9 +1,14 @@
 package com.dj.insulink.feature.meals.ui
 
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CalendarToday
 import androidx.compose.material.icons.filled.Schedule
@@ -12,7 +17,6 @@ import androidx.compose.material3.DatePicker
 import androidx.compose.material3.DatePickerDialog
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.InputChip
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
@@ -26,8 +30,9 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.font.FontWeight
 import com.dj.insulink.R
 import com.dj.insulink.core.ui.theme.InsulinkTheme
 import java.text.SimpleDateFormat
@@ -39,7 +44,8 @@ import java.util.Locale
 @Composable
 fun DateTimeInput(
     selectedTimestamp: Long,
-    onTimestampSelected: (Long) -> Unit
+    onTimestampSelected: (Long) -> Unit,
+    modifier: Modifier = Modifier
 ) {
     var showDatePicker by remember { mutableStateOf(false) }
     var showTimePicker by remember { mutableStateOf(false) }
@@ -49,27 +55,67 @@ fun DateTimeInput(
     val timeFormatter = remember { SimpleDateFormat("HH:mm", Locale.getDefault()) }
     val date = Date(selectedTimestamp)
 
+    val chipShape = RoundedCornerShape(InsulinkTheme.dimens.commonButtonRadius12)
+
     Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
+        modifier = modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.spacedBy(InsulinkTheme.dimens.commonSpacing12)
     ) {
-        InputChip(
-            selected = false,
-            onClick = { showDatePicker = true },
-            label = { Text(text = dateFormatter.format(date), textAlign = TextAlign.Center) },
-            leadingIcon = { Icon(Icons.Filled.CalendarToday, contentDescription = "") },
-            modifier = Modifier.weight(1f)
-        )
+        Row(
+            modifier = Modifier
+                .weight(1f)
+                .height(InsulinkTheme.dimens.commonButtonHeight50)
+                .clip(chipShape)
+                .border(
+                    InsulinkTheme.dimens.commonButtonBorder1,
+                    MaterialTheme.colorScheme.outline,
+                    chipShape
+                )
+                .clickable { showDatePicker = true },
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.Center
+        ) {
+            Icon(
+                Icons.Filled.CalendarToday,
+                contentDescription = "",
+                modifier = Modifier.size(InsulinkTheme.dimens.commonIconSize16),
+                tint = MaterialTheme.colorScheme.primary
+            )
+            Spacer(Modifier.size(InsulinkTheme.dimens.commonSpacing8))
+            Text(
+                text = dateFormatter.format(date),
+                style = MaterialTheme.typography.bodyLarge,
+                fontWeight = FontWeight.Medium
+            )
+        }
 
-        Spacer(Modifier.size(InsulinkTheme.dimens.commonSpacing8))
-
-        InputChip(
-            selected = false,
-            onClick = { showTimePicker = true },
-            label = { Text(text = timeFormatter.format(date), textAlign = TextAlign.Center) },
-            leadingIcon = { Icon(Icons.Filled.Schedule, contentDescription = "") },
-            modifier = Modifier.weight(1f)
-        )
+        Row(
+            modifier = Modifier
+                .weight(1f)
+                .height(InsulinkTheme.dimens.commonButtonHeight50)
+                .clip(chipShape)
+                .border(
+                    InsulinkTheme.dimens.commonButtonBorder1,
+                    MaterialTheme.colorScheme.outline,
+                    chipShape
+                )
+                .clickable { showTimePicker = true },
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.Center
+        ) {
+            Icon(
+                Icons.Filled.Schedule,
+                contentDescription = "",
+                modifier = Modifier.size(InsulinkTheme.dimens.commonIconSize16),
+                tint = MaterialTheme.colorScheme.primary
+            )
+            Spacer(Modifier.size(InsulinkTheme.dimens.commonSpacing8))
+            Text(
+                text = timeFormatter.format(date),
+                style = MaterialTheme.typography.bodyLarge,
+                fontWeight = FontWeight.Medium
+            )
+        }
     }
 
     if (showDatePicker) {
